+++ patches/Accessibility/strings_fr.cpp
@@ -444,6 +444,7 @@ const char* Get(Id id) {
         case Id::WorkbenchSlotPeekEmpty:     return "Emplacement vide, aucune am\xE9lioration install\xE9""e";
         case Id::WorkbenchFmtSlotItem:       return "%s, occup\xE9 par %s";
         case Id::WorkbenchPickerInstalled:   return "install\xE9";
+        case Id::WorkbenchPickerNone:        return "Aucun objet";
         case Id::WorkbenchModeCreate:        return "Mode fabrication";
         case Id::WorkbenchModeBreakdown:     return "Mode d\xE9montage";
         case Id::FmtWorkbenchUpgradeTitle:   return "Am\xE9lioration : %s";